added test dir, refactored src dir, split dev deps into extras
