+++ README.md
@@ -1,3 +1,25 @@
 # elo-cli
 
-Calculate ELO ratings in your terminal.
+[![Build Status](https://api.travis-ci.com/bradleycwojcik/elo-cli.svg?branch=main)](https://travis-ci.com/github/bradleycwojcik/elo-cli)
+[![codecov](https://codecov.io/gh/bradleycwojcik/elo-cli/branch/main/graph/badge.svg)](https://codecov.io/gh/bradleycwojcik/elo-cli)
+[![Docs](https://img.shields.io/website?down_message=down&label=docs&up_message=online&url=https%3A%2F%2Fbradleycwojcik.github.io%2Felo-cli%2F)](https://bradleycwojcik.github.io/elo-cli/)
+[![PyPI](https://img.shields.io/pypi/v/elo-cli)](https://pypi.org/project/elo-cli/)
+![PyPI - Downloads](https://img.shields.io/pypi/dm/elo-cli)
+
+> Calculate ELO ratings in your terminal.
+
+elo-cli is a tool for calculating and storing elo ratings and statistics for two sided match-based games.
+
+## Install
+
+Install to a global executable isolated environment with [`pipx`](https://pypa.github.io/pipx/) (Recommended).
+
+```zsh
+pipx install elo-cli
+```
+
+Or install to your active python instance (Python 3.8+ required).
+
+```zsh
+pip install elo-cli
+```
