+++ elo/command/series_grp/series.py
