+++ elo/model/base.py
