+++ elo/model/player.py
