+++ tests/test_elo.py
@@ -1,0 +1,15 @@
+from click.testing import CliRunner
+
+from elo.elo import cli
+
+
+def test_version():
+    runner = CliRunner()
+    result = runner.invoke(cli, ["--version"])
+    assert result.exit_code == 0
+
+
+def test_help():
+    runner = CliRunner()
+    result = runner.invoke(cli, ["--help"])
+    assert result.exit_code == 0
